requirements.txt
biopython==1.79
cycler==0.11.0
fonttools==4.28.2
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
setuptools-scm==6.3.2
six==1.16.0
tomli==1.2.2
